refactor: Centralize Book and Lesson type definitions
Centralized the Book and Lesson interfaces into `src/data/lessons.ts`.
Updated `src/data/book1.ts`, `src/data/book2.ts`, and `src/data/book3.ts` to import these interfaces from the new central location.
This resolves the issue where `book3.ts` was incorrectly referencing `book1.ts` for these types and improves maintainability.

Completed TODO: Refactor Book and Lesson type definitions to `src/data/lessons.ts`.
