+++ TODO.md
@@ -71,6 +71,7 @@
 - [x] Update README.md to reflect current application state. (commit: 2ef4d9f)
 - [x] Implement page transitions and mobile sidebar animations. (5dbd11a)
 - [x] Remove all transitions from the project. (commit: cb687ab)
+- [x] Refactor Book and Lesson type definitions to `src/data/lessons.ts`.
 - [x] Implement changelog feature based on commit history. (5adfafb)
  - [x] Add 'Return to Homepage' link to sidebars. (acc5bd0)
   - [x] Fix missing layout (sidebar and homepage button) on Changelog page. (de1b501)
