+++ src/data/book3.ts
@@ -1,4 +1,4 @@
-import { Book, Lesson } from './book1';
+import { Book, Lesson } from './lessons';
 
 export const book3Data: Book = {
   id: 'book3',
